Eliminar roles corregido

+++ Server/src/build/controllers/rolesController.js
@@ -59,10 +59,18 @@ class RolesController {
             // Obtener nombre_rol del rol que se va a eliminar
             const respNombreRol = yield database_1.default.query(`SELECT nombre_rol FROM roles WHERE id_rol = ${id}`);
             const nombreRol = (_a = respNombreRol[0]) === null || _a === void 0 ? void 0 : _a.nombre_rol;
-            // Eliminar ofertas laborales donde aparezca el nombre del rol en el campo puesto
-            yield database_1.default.query(`DELETE FROM ofertalaboral WHERE puesto = '${nombreRol}'`);
-            // Eliminar usuarios con el mismo id_rol
-            yield database_1.default.query(`DELETE FROM usuarios WHERE id_rol = ${id}`);
+            //Buscar si existen ofertas laborales asociadas al rol
+            const ofertasLaborales = yield database_1.default.query(`SELECT * FROM ofertalaboral, roles WHERE "${nombreRol}" = ofertalaboral.puesto LIMIT 1`);
+            if (ofertasLaborales[0] != null) {
+                res.json({ exito: -1, message: 'No se puede eliminar el rol porque existen ofertas laborales asociadas a este rol' });
+                return;
+            }
+            //Buscamos si hay usuarios con ese rol
+            const usuarios = yield database_1.default.query(`SELECT * FROM usuarios WHERE id_rol = ${id} LIMIT 1`);
+            if (usuarios[0] != null) {
+                res.json({ exito: -1, message: 'No se puede eliminar el rol porque existen usuarios asociados a este rol' });
+                return;
+            }
             // Eliminar el rol
             const resp = yield database_1.default.query(`DELETE FROM roles WHERE id_rol = ${id}`);
             res.json(resp);
